Fixed false positives in the culling system.
This fixes numerous false positives coming out of the culling system.

AABB checks are now a full separating-axis check against the frustum, with the points of the frustum being compared to the planes of the box just as the points of the box were being compared to the planes of the frustum. This fixes large objects behind the camera not being culled correctly.

A new class, ConvexShape, was added to provide a full set of planes+points for arbitrary convex meshes, and also because passing around only planes was no longer sufficient for culling operations. This class will calculate intersection points for planes and store them.

A lot of the culling interface has been given a ConvexShape version to use instead of taking an array of planes. The remaining functions in the old interface now create a ConvexShape internally based on the planes given.

Some systems that used frustums that were (sometimes mistakenly?) unbounded on one or more side have been modified to be fully enclosed. This change fixed unbounded culling by the omni light dual paraboloid shadowmap culling.